move test to proxy-providers.spec.ts
